Split GUI fixture helpers
